src/requirements.txt
Flask==2.2.3
minio==7.1.12
PyJWT==2.6.0